Improving Stack Analysis
How it works with this update:
Analyze a given method. If a type can't be determined it'll just use a
descriptor as a placeholder (No values). If a method is called and it
returns a type it will be emulated for the final return value.

It's all a bit dirty right now and the emulation isn't perfect. I'm
using a string encryption sample. It's close but not exactly working.
Gonna sleep on it and hopefully figure out whats wrong soon.

I also made a logger class. It's pretty bad but better than scattered
around System.out.println's. I'll eventually get around to replacing all
those calls...